dgx/resources/requirements.txt
Flask
Flask-Cors
flasgger
Flask-SocketIO
pymongo
eventlet
requests
Flask-JWT
redis
cerberus